+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Adding a new version? You'll need three changes:
 * Add the diff link, like "[2.7.0]: https://github.com/kong/kubernetes-ingress-controller/compare/v1.2.2...v1.2.3".
   This is all the way at the bottom. It's the thing we always forget.
 --->
+ - [3.3.0](#330)
  - [3.2.4](#324)
  - [3.2.3](#323)
  - [3.2.2](#322)
@@ -96,7 +97,9 @@ Adding a new version? You'll need three changes:
  - [0.0.5](#005)
  - [0.0.4 and prior](#004-and-prior)
 
-## Unreleased
+## [3.3.0]
+
+> Release date: 2024-08-26
 
 ### Added